+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+### Added
+
+- Add click-cast support for `Boss5TargetFrame` [#45]
+
 ## [0.15.1] - 2021-01-21
 
 ### Added
@@ -537,6 +543,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 [0.4.0]: https://github.com/Snakybo/Clicked/releases/tag/0.4.0
 [0.3.0]: https://github.com/Snakybo/Clicked/releases/tag/0.3.0
 
+[#45]: https://github.com/Snakybo/Clicked/issues/45
 [#41]: https://github.com/Snakybo/Clicked/issues/41
 [#40]: https://github.com/Snakybo/Clicked/issues/40
 [#39]: https://github.com/Snakybo/Clicked/issues/39